+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.0 under development
 
 - Initial release.
+- New #263: Add php attribute SkipCollection to skip class properties on collection